add Access controller add image to posts

+++ app/Http/Controllers/PostsController.php
@@ -3,9 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 class PostsController extends Controller
 {
+
+
+      /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth' , ['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -38,14 +51,32 @@ class PostsController extends Controller
         // Check that the request data is valid
         $this->validate($request , [
             'title'=>'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file
+        if($request->hasFile('cover_image')){
+            // Get File name extension
+            $fileNameWithEXT = $request->file('cover_image')->getClientOriginalName();
+            // file name
+            $fileName = pathinfo($fileNameWithEXT,PATHINFO_FILENAME);
+            // get extension
+            $extension = $request->file('cover_image')->guessClientExtension();
+            // file to store
+            $fileToStore = $fileName.'_'.time().'.'.$extension;
+            //Upload Image
+            $path  = $request->file('cover_image')->storeAs('public/cover_images' , $fileToStore);
+        }else {
+            $fileToStore = 'noImage.jpg';
+        }
 
         //Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
         $post->user_id = auth()->user()->id ;
+        $post->cover_image = $fileToStore ;
         $post->save();
         return redirect('/posts')->with('success','Post Created Successfully');
     }
@@ -71,6 +102,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check if the user own the post
+        if(auth()->user()->id !== $post->user_id){
+             return redirect('/posts')->with('error' , 'UnAuthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -86,14 +121,34 @@ class PostsController extends Controller
         // Check that the request data is valid
         $this->validate($request , [
             'title'=>'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file
+        if($request->hasFile('cover_image')){
+            // Get File name extension
+            $fileNameWithEXT = $request->file('cover_image')->getClientOriginalName();
+            // file name
+            $fileName = pathinfo($fileNameWithEXT,PATHINFO_FILENAME);
+            // get extension
+            $extension = $request->file('cover_image')->guessClientExtension();
+            // file to store
+            $fileToStore = $fileName.'_'.time().'.'.$extension;
+            //Upload Image
+            $path  = $request->file('cover_image')->storeAs('public/cover_images' , $fileToStore);
+        }else {
+            $fileToStore = 'noImage.jpg';
+        }
 
         //Create Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
-       
+        if($request->hasFile('cover_image')){
+            Storage::delete('public/cover_images/' . $post->cover_image);
+            $post->cover_image = $fileToStore;
+        }
         $post->save();
         return redirect('/posts')->with('success','Post Updated Successfully');
     }
@@ -107,6 +162,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error' , ' Page');
+       }
+       if($post->cover_image !='noImage.jpg'){
+           // Delete Image
+           Storage::delete('public/cover_images/'.$post->cover_image);
+       }
         $post->delete();
         return redirect('/posts')->with('success','Post Deleted Successfully');
     }
